Reorganize bundle and database commands into separate namespaces.
