+++ src/tools/wix/Bind/Bundles/OrderPackagesAndRollbackBoundariesCommand.cs
@@ -7,7 +7,7 @@
 // </copyright>
 //-------------------------------------------------------------------------------------------------
 
-namespace WixToolset.Bind
+namespace WixToolset.Bind.Bundles
 {
     using System;
     using System.Collections.Generic;
