+++ src/tools/wix/Bind/Bundles/PackageFacade.cs
@@ -7,7 +7,7 @@
 // </copyright>
 //-------------------------------------------------------------------------------------------------
 
-namespace WixToolset.Bind
+namespace WixToolset.Bind.Bundles
 {
     using WixToolset.Data.Rows;
