+++ src/tools/wix/Bind/Bundles/ProcessMsuPackage.cs
@@ -7,7 +7,7 @@
 // </copyright>
 //-------------------------------------------------------------------------------------------------
 
-namespace WixToolset.Bind
+namespace WixToolset.Bind.Bundles
 {
     using System;
     using WixToolset.Data;
