+++ src/tools/wix/Bind/Databases/BindSummaryInfoCommand.cs
@@ -7,13 +7,10 @@
 // </copyright>
 //-------------------------------------------------------------------------------------------------
 
-namespace WixToolset.Bind
+namespace WixToolset.Bind.Databases
 {
     using System;
-    using System.Collections.Generic;
     using System.Globalization;
-    using System.Linq;
-    using System.Text;
     using WixToolset.Data;
 
     /// <summary>
