+++ src/tools/wix/Unbinder.cs
@@ -5,10 +5,6 @@
 //   The license and further copyright text can be found in the file
 //   LICENSE.TXT at the root directory of the distribution.
 // </copyright>
-// 
-// <summary>
-// Unbinder core of the WiX toolset.
-// </summary>
 //-------------------------------------------------------------------------------------------------
 
 namespace WixToolset
@@ -24,6 +20,7 @@ namespace WixToolset
     using System.Linq;
     using System.Text.RegularExpressions;
     using WixToolset.Bind;
+    using WixToolset.Bind.Bundles;
     using WixToolset.Cab;
     using WixToolset.Data;
     using WixToolset.Data.Rows;
