+++ src/tools/wix/Wix.csproj
@@ -18,53 +18,66 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="AssemblyInfo.cs" />
-    <Compile Include="Bind\AutomaticallySlipstreamPatchesCommand.cs" />
-    <Compile Include="Bind\AssignMediaCommand.cs" />
+    <Compile Include="Bind\Bundles\AutomaticallySlipstreamPatchesCommand.cs" />
+    <Compile Include="Bind\Databases\AssignMediaCommand.cs" />
     <Compile Include="Binder.cs" />
     <Compile Include="BinderCore.cs" />
     <Compile Include="BinderFileManagerCore.cs" />
     <Compile Include="BinderFileManager.cs" />
+    <Compile Include="CommandLine.cs" />
+    <Compile Include="Common.cs" />
+    <Compile Include="Compiler.cs" />
+    <Compile Include="CompilerCore.cs" />
+    <Compile Include="Differ.cs" />
+    <Compile Include="ExtensionManager.cs" />
+    <Compile Include="IfDefEventHandler.cs" />
+    <Compile Include="Inscriber.cs" />
+    <Compile Include="InspectorCore.cs" />
+    <Compile Include="ICommand.cs" />
+    <Compile Include="Localizer.cs" />
+    <Compile Include="Patch.cs" />
+    <Compile Include="PatchTransform.cs" />
+    <Compile Include="ResolvedVariableEventHandler.cs" />
+    <Compile Include="Uuid.cs" />
     <Compile Include="Bind\BindBundleCommand.cs" />
     <Compile Include="Bind\BindDatabaseCommand.cs" />
-    <Compile Include="Bind\BindSummaryInfoCommand.cs" />
     <Compile Include="Bind\BindTransformCommand.cs" />
-    <Compile Include="Bind\PackageFacade.cs" />
-    <Compile Include="Bind\ConfigurationCallback.cs" />
-    <Compile Include="Bind\CopyTransformDataCommand.cs" />
-    <Compile Include="Bind\CreateBootstrapperApplicationManifestCommand.cs" />
-    <Compile Include="Bind\CreateBurnManifestCommand.cs" />
-    <Compile Include="Bind\CreateContainerCommand.cs" />
-    <Compile Include="Bind\CreateCabinetsCommand.cs" />
     <Compile Include="Bind\DelayedField.cs" />
+    <Compile Include="Bind\ExtractEmbeddedFiles.cs" />
     <Compile Include="Bind\ExtractEmbeddedFilesCommand.cs" />
+    <Compile Include="Bind\FileTransfer.cs" />
     <Compile Include="Bind\GenerateDatabaseCommand.cs" />
-    <Compile Include="Bind\GetPackageFacadesCommand.cs" />
-    <Compile Include="Bind\OrderPackagesAndRollbackBoundariesCommand.cs" />
-    <Compile Include="Bind\ProcessMspPackage.cs" />
-    <Compile Include="Bind\ProcessMsuPackage.cs" />
-    <Compile Include="Bind\ProcessExePackage.cs" />
-    <Compile Include="Bind\ProcessMsiPackage.cs" />
-    <Compile Include="Bind\ProcessPayloads.cs" />
     <Compile Include="Bind\ResolveDelayedFieldsCommand.cs" />
     <Compile Include="Bind\ResolveFieldsCommand.cs" />
     <Compile Include="Bind\TransferFilesCommand.cs" />
-    <Compile Include="Bind\VerifyPayloadsWithCatalogCommand.cs" />
-    <Compile Include="Bind\WixPatchSymbolPathsComparer.cs" />
-    <Compile Include="Bind\BurnCommon.cs" />
-    <Compile Include="Bind\BurnReader.cs" />
-    <Compile Include="Bind\BurnWriter.cs" />
-    <Compile Include="Bind\CabinetBuilder.cs" />
-    <Compile Include="Bind\CabinetWorkItem.cs" />
+    <Compile Include="Bind\Bundles\BurnCommon.cs" />
+    <Compile Include="Bind\Bundles\BurnReader.cs" />
+    <Compile Include="Bind\Bundles\BurnWriter.cs" />
+    <Compile Include="Bind\Bundles\PackageFacade.cs" />
+    <Compile Include="Bind\Bundles\CreateBootstrapperApplicationManifestCommand.cs" />
+    <Compile Include="Bind\Bundles\CreateBurnManifestCommand.cs" />
+    <Compile Include="Bind\Bundles\CreateContainerCommand.cs" />
+    <Compile Include="Bind\Bundles\GetPackageFacadesCommand.cs" />
+    <Compile Include="Bind\Bundles\OrderPackagesAndRollbackBoundariesCommand.cs" />
+    <Compile Include="Bind\Bundles\ProcessMspPackage.cs" />
+    <Compile Include="Bind\Bundles\ProcessMsuPackage.cs" />
+    <Compile Include="Bind\Bundles\ProcessExePackage.cs" />
+    <Compile Include="Bind\Bundles\ProcessMsiPackage.cs" />
+    <Compile Include="Bind\Bundles\ProcessPayloads.cs" />
+    <Compile Include="Bind\Bundles\VerifyPayloadsWithCatalogCommand.cs" />
+    <Compile Include="Bind\Databases\BindSummaryInfoCommand.cs" />
+    <Compile Include="Bind\Databases\CabinetBuilder.cs" />
+    <Compile Include="Bind\Databases\CabinetWorkItem.cs" />
+    <Compile Include="Bind\Databases\CreateCabinetsCommand.cs" />
+    <Compile Include="Bind\Databases\ConfigurationCallback.cs" />
+    <Compile Include="Bind\Databases\CopyTransformDataCommand.cs" />
+    <Compile Include="Bind\Databases\WixPatchSymbolPathsComparer.cs" />
     <Compile Include="Cab\CabinetFileInfo.cs" />
     <Compile Include="Cab\Interop\CabInterop.cs" />
     <Compile Include="Cab\WixCreateCab.cs" />
     <Compile Include="Cab\WixEnumerateCab.cs" />
     <Compile Include="Cab\WixExtractCab.cs" />
     <Compile Include="Clr\Interop\ClrInterop.cs" />
-    <Compile Include="CommandLine.cs" />
-    <Compile Include="Common.cs" />
-    <Compile Include="Compiler.cs" />
-    <Compile Include="CompilerCore.cs" />
     <Compile Include="Extensibility\AssemblyDefaultHeatExtensionAttribute.cs" />
     <Compile Include="Extensibility\HarvesterExtension.cs" />
     <Compile Include="Extensibility\HeatExtension.cs" />
@@ -72,23 +85,10 @@
     <Compile Include="Extensibility\IHeatCore.cs" />
     <Compile Include="Extensibility\MutatorExtension.cs" />
     <Compile Include="Extensibility\ValidatorExtension.cs" />
-    <Compile Include="ExtensionManager.cs" />
-    <Compile Include="Bind\ExtractEmbeddedFiles.cs" />
-    <Compile Include="Bind\FileTransfer.cs" />
     <Compile Include="Link\FindEntrySectionAndLoadSymbolsCommand.cs" />
-    <Compile Include="IfDefEventHandler.cs" />
-    <Compile Include="Inscriber.cs" />
-    <Compile Include="InspectorCore.cs" />
-    <Compile Include="ICommand.cs" />
     <Compile Include="Link\ReportConflictingSymbolsCommand.cs" />
-    <Compile Include="Localizer.cs" />
-    <Compile Include="Patch.cs" />
-    <Compile Include="PatchAPI\PatchInterop.cs" />
-    <Compile Include="PatchTransform.cs" />
-    <Compile Include="ResolvedVariableEventHandler.cs" />
     <Compile Include="Link\ResolveReferencesCommand.cs" />
-    <Compile Include="Uuid.cs" />
-    <Compile Include="Differ.cs" />
+    <Compile Include="PatchAPI\PatchInterop.cs" />
     <Compile Include="Ole32\Storage.cs" />
     <Compile Include="WixDistributionSpecificStrings.Designer.cs">
       <AutoGen>True</AutoGen>
